+++ IntentTodo.xcodeproj/xcuserdata/touyou.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -14,10 +14,20 @@
 			<key>orderHint</key>
 			<integer>1</integer>
 		</dict>
+		<key>IntentTodoUITest.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>8</integer>
+		</dict>
 		<key>IntentTodoWatchApp Watch App.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>3</integer>
+		</dict>
+		<key>IntentTodoWatchAppUITest.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>9</integer>
 		</dict>
 		<key>IntentTodoWidgetExtension.xcscheme_^#shared#^_</key>
 		<dict>
@@ -43,6 +53,16 @@
 			<true/>
 		</dict>
 		<key>46AFBEFC2F2D6E5A00444306</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>46AFC0F22F34F72800444306</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>46AFC1062F34F75C00444306</key>
 		<dict>
 			<key>primary</key>
 			<true/>
