+++ IntentTodoLiveActivity/Intents/SnoozeTodoIntent.swift
@@ -1,8 +1,8 @@
 //
-//  LiveActivityIntents.swift
+//  SnoozeTodoIntent.swift
 //  IntentTodoLiveActivity
 //
-//  Intents for Live Activity actions.
+//  Intent to snooze a todo deadline from Live Activity.
 //
 
 import ActivityKit
@@ -11,65 +11,6 @@ import Domain
 import Repository
 import SwiftData
 import WidgetKit
-
-// MARK: - Model Container
-
-/// Uses SharedModelContainer for data sharing with the main app.
-/// Requires App Group to be configured in Xcode.
-let liveActivityModelContainer: ModelContainer = {
-    // swiftlint:disable:next force_try
-    return try! SharedModelContainer.createContainer()
-}()
-
-// MARK: - Complete Todo Intent
-
-/// Intent to complete a todo from Live Activity.
-struct CompleteTodoFromActivityIntent: LiveActivityIntent {
-    static var title: LocalizedStringResource = "Complete Todo"
-
-    @Parameter(title: "Todo ID")
-    var todoId: String
-
-    init() {
-        self.todoId = ""
-    }
-
-    init(todoId: String) {
-        self.todoId = todoId
-    }
-
-    @MainActor
-    func perform() async throws -> some IntentResult {
-        guard let uuid = UUID(uuidString: todoId) else {
-            return .result()
-        }
-
-        let repository = SwiftDataTodoRepository(modelContext: liveActivityModelContainer.mainContext)
-        if let todo = try await repository.fetch(by: uuid) {
-            todo.isCompleted = true
-            try await repository.update(todo)
-
-            // Reload widgets to reflect the change
-            WidgetCenter.shared.reloadAllTimelines()
-
-            // End the Live Activity
-            await endLiveActivity(for: todoId)
-        }
-
-        return .result()
-    }
-
-    @MainActor
-    private func endLiveActivity(for todoId: String) async {
-        for activity in Activity<TodoDeadlineActivityAttributes>.activities {
-            if activity.attributes.todoId == todoId {
-                await activity.end(dismissalPolicy: .immediate)
-            }
-        }
-    }
-}
-
-// MARK: - Snooze Todo Intent
 
 /// Intent to snooze a todo deadline from Live Activity.
 struct SnoozeTodoIntent: LiveActivityIntent {
